Fix so that up to 4 client messages can be processed concurrently
The call to `tokio::threadpool::blocking()` will block all other futures
on the task as well.  Therefore, sub-tasks are spawned in the client so
message processing does not block the main client task.

A handle to the spawned future is obtained and returned, otherwise
sub-tasks will be spawned until the `max_blocking` thread count is
reached.